setup react project
