+++ src/App.js
@@ -1,0 +1,7 @@
+function App() {
+  return (
+    <div className="App" />
+  );
+}
+
+export default App;
